fix(897): Fixed Raw Body Parsing in intial loading
